Initial commit
Rudimentary implementation of converting an image file to a map item in minecraft.

Uses a basic average HSV comparison method to convert each pixel of a given image into one of 177 colours available on a minecraft map.

The code then writes minecraft commands to place corresponding blocks in a set 128x256x128 area in the world. The player can then approach and display the image using a map.