Updated to use CIE-L*ab colour space
A much more complex and accurate colour comparison algorithm has been implemented. Colours are no longer stored as HSV and instead use the CIE-L*ab model.

While results are more accurate, the execution time for the algorithm to run is very high.

There is a bug which seems to cause the lightness of pixels to be inverted