Fixed slope error
Blocks were previously being placed at the incorrect y-position, causing the resulting image to appear blocky. This has been fixed.

Currently, in the rare case an entire column of pixels all have a positive or negative slope value, the bottom pixel will be grey as a block will be attempted to be placed outside of the world.